// types/fetchedHackathonData.ts
export interface fetchedHackathonData {
  image_url: string;
  title: string;
  days_left: number;
  online: boolean;
  prize_amount: string;
  participants: number;
  host_name: string;
  host_logo_url: string;
  submission_period: string;
  managed_by: string;
  managed_by_logo_url: string;
  themes: string[];
}